+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Renamed packages "core" to "api", "contrib" to "impl", "utils" to "util"
 - Reorganized the project structure (moved all providers into the "impl" package, moved some classes/traits)
 - Add request handlers
+- Add request providers in combination with HTTP basic auth provider
 
 ## 1.0 (2014-06-12)